text_to_video_agent/pyproject.toml
[project]
name = "text-to-video-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi_poe>=0.0.36",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "openai>=2.7.2",
]
